# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ARAL System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /teachers:
    get:
      operationId: getTeachers
      tags: [teachers]
      summary: Get teachers for this school
      responses:
        "200":
          description: Teachers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Teacher"
    post:
      operationId: createTeacher
      tags: [teachers]
      summary: Add a teacher to a grade level
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherInput"
      responses:
        "201":
          description: Teacher added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
  /teachers/{teacherId}:
    get:
      operationId: getTeacher
      tags: [teachers]
      summary: Get teacher details
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teacher details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Teacher"
    delete:
      operationId: deleteTeacher
      tags: [teachers]
      summary: Delete a teacher and connected records
      parameters:
        - name: teacherId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teacher deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /teachers/profile:
    get:
      operationId: getTeacherProfile
      tags: [teachers]
      summary: Get current teacher's profile
      responses:
        "200":
          description: Teacher profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherProfile"
    post:
      operationId: saveTeacherProfile
      tags: [teachers]
      summary: Save teacher profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeacherProfileInput"
      responses:
        "200":
          description: Profile saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeacherProfile"

  # Learners
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message
...
    Teacher:
      type: object
      properties:
        id:
          type: string
        schoolId:
          type: string
        gradeLevelId:
          type: string
        gradeLevelName:
          type: string
        firstName:
          type: string
        middleName:
          type: ["string", "null"]
        lastName:
          type: string
        pin:
          type: ["string", "null"]
        profileComplete:
          type: boolean
        isActive:
          type: boolean
        learnersCount:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - schoolId
        - gradeLevelId
        - gradeLevelName
        - firstName
        - lastName
        - profileComplete
        - isActive
        - learnersCount
        - createdAt
    TeacherInput:
      type: object
      properties:
        firstName:
          type: string
        middleName:
          type: string
        lastName:
          type: string
        gradeLevelId:
          type: string
        pin:
          type: string
          pattern: ^[0-9]{4,12}$
      required:
        - firstName
        - lastName
        - gradeLevelId
        - pin
    TeacherProfile:
      type: object
      properties:
        id:
          type: string
        teacherId:
          type: string
        name:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        sex:
          type: ["string", "null"]
        dateOfBirth:
          type: ["string", "null"]
        designation:
          type: string
        designationOther:
          type: ["string", "null"]
        position:
          type: string
        email:
          type: string
        district:
          type: ["string", "null"]
        division:
          type: ["string", "null"]
        schoolYear:
          type: ["string", "null"]
        yearsInService:
          type: string
        highestEducationalAttainment:
          type: string
        fieldOfSpecialization:
          type: string
        fieldOfSpecializationOther:
          type: ["string", "null"]
        currentGradeLevel:
          type: string
        contactNumber:
          type: string
        mostSubjectHandled:
          type: string
        trainingsAttended:
          type: array
          items:
            type: string
        literacyTrainingAttended:
          type: ["string", "null"]
        readingTrainingsAttended:
          type: array
          items:
            type: string
        englishTrainingAttended:
          type: ["string", "null"]
        englishTrainingsAttended:
          type: array
          items:
            type: string
        highestTrainingLevel:
          type: ["string", "null"]
        isComplete:
          type: boolean
      required:
        - id
        - teacherId
        - designation
        - position
        - email
        - yearsInService
        - highestEducationalAttainment
        - fieldOfSpecialization
        - currentGradeLevel
        - contactNumber
        - mostSubjectHandled
        - trainingsAttended
        - readingTrainingsAttended
        - englishTrainingsAttended
        - isComplete
    TeacherProfileInput:
      type: object
      properties:
        name:
          type: string
        age:
          type: integer
        sex:
          type: string
        dateOfBirth:
          type: string
        designation:
          type: string
        designationOther:
          type: string
        position:
          type: string
        email:
          type: string
        district:
          type: string
        division:
          type: string
        schoolYear:
          type: string
        yearsInService:
          type: string
        highestEducationalAttainment:
          type: string
        fieldOfSpecialization:
          type: string
        fieldOfSpecializationOther:
          type: string
        currentGradeLevel:
          type: string
        contactNumber:
          type: string
        mostSubjectHandled:
          type: string
        trainingsAttended:
          type: array
          items:
            type: string
        literacyTrainingAttended:
          type: string
        readingTrainingsAttended:
          type: array
          items:
            type: string
        englishTrainingAttended:
          type: string
        englishTrainingsAttended:
          type: array
          items:
            type: string
        highestTrainingLevel:
          type: string
      required:
        - designation
        - position
        - email
        - yearsInService
        - highestEducationalAttainment
        - fieldOfSpecialization
        - currentGradeLevel
        - contactNumber
        - mostSubjectHandled
        - literacyTrainingAttended
        - englishTrainingAttended
        - highestTrainingLevel
...