# lib/api-zod/src/generated/types/teacherInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARAL System API
 * OpenAPI spec version: 0.1.0
 */

export interface TeacherInput {
  firstName: string;
  middleName?: string;
  lastName: string;
  gradeLevelId: string;
  /** @pattern ^[0-9]{4,12}$ */
  pin: string;
}