# lib/api-zod/src/generated/types/teacherProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARAL System API
 * OpenAPI spec version: 0.1.0
 */

export interface TeacherProfile {
  id: string;
  teacherId: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  sex?: string | null;
  /** @nullable */
  dateOfBirth?: string | null;
  designation: string;
  /** @nullable */
  designationOther?: string | null;
  position: string;
  email: string;
  /** @nullable */
  district?: string | null;
  /** @nullable */
  division?: string | null;
  /** @nullable */
  schoolYear?: string | null;
  yearsInService: string;
  highestEducationalAttainment: string;
  fieldOfSpecialization: string;
  /** @nullable */
  fieldOfSpecializationOther?: string | null;
  currentGradeLevel: string;
  contactNumber: string;
  mostSubjectHandled: string;
  trainingsAttended: string[];
  /** @nullable */
  literacyTrainingAttended?: string | null;
  readingTrainingsAttended: string[];
  /** @nullable */
  englishTrainingAttended?: string | null;
  englishTrainingsAttended: string[];
  /** @nullable */
  highestTrainingLevel?: string | null;
  isComplete: boolean;
}