# lib/api-zod/src/generated/types/teacherProfileInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ARAL System API
 * OpenAPI spec version: 0.1.0
 */

export interface TeacherProfileInput {
  name?: string;
  age?: number;
  sex?: string;
  dateOfBirth?: string;
  designation: string;
  designationOther?: string;
  position: string;
  email: string;
  district?: string;
  division?: string;
  schoolYear?: string;
  yearsInService: string;
  highestEducationalAttainment: string;
  fieldOfSpecialization: string;
  fieldOfSpecializationOther?: string;
  currentGradeLevel: string;
  contactNumber: string;
  mostSubjectHandled: string;
  trainingsAttended?: string[];
  literacyTrainingAttended: string;
  readingTrainingsAttended?: string[];
  englishTrainingAttended: string;
  englishTrainingsAttended?: string[];
  highestTrainingLevel: string;
}